pyproject.toml
[project]
name = "api-meta"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apache-airflow>=2.10.5",
    "ipykernel>=6.29.5",
    "pandas>=2.2.3",
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
]
